Added lazysmp

+++ rust_mbu/rust_engine/engine/src/engine/eval/mod.rs
@@ -4,13 +4,13 @@ pub trait Evaluate {
     fn evaluate(&self, position: &Position) -> f64;
 }
 
-const PIECE_VALUES: [isize; 6] = [100, 300, 350, 500, 900, 10000];
+const PIECE_VALUES: [i32; 6] = [100, 300, 350, 500, 900, 10000];
 
-pub fn evaluate(position: &Position) -> isize {
+pub fn evaluate(position: &Position) -> i32 {
     let mut score = 0;
     for piece in Piece::all() {
-        let white_count = position.pieces[Color::White as usize][piece as usize].count() as isize;
-        let black_count = position.pieces[Color::Black as usize][piece as usize].count() as isize;
+        let white_count = position.pieces[Color::White as usize][piece as usize].count() as i32;
+        let black_count = position.pieces[Color::Black as usize][piece as usize].count() as i32;
 
         score += PIECE_VALUES[piece as usize] * (white_count - black_count);
     }
